Remove implicit-int-conversion warnings.
BUG=webrtc:1348, webrtc:261
R=stefan@webrtc.org

Review URL: https://codereview.webrtc.org/1184443005.

Cr-Commit-Position: refs/heads/master@{#9464}

+++ webrtc/video_engine/overuse_frame_detector.cc
@@ -250,7 +250,7 @@ class OveruseFrameDetector::FrameQueue {
   }
 
   void Reset() { frame_times_.clear(); }
-  int NumFrames() const { return frame_times_.size(); }
+  int NumFrames() const { return static_cast<int>(frame_times_.size()); }
   int last_processing_time_ms() const { return last_processing_time_ms_; }
 
  private:
